requirements.txt
azure-core==1.35.0
azure-storage-blob==12.26.0
boto3==1.39.7
botocore==1.39.7
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
et_xmlfile==2.0.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
lxml==6.0.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
pycparser==2.22
pyotp==2.9.0
pyPreservica==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
